requirements.txt
grpcio
grpcio-tools
sqlalchemy
mysql-connector-python
yfinance
